Some basic code optimizations
More specifically:
* Add gitignore
* Replace if chains with switch-case statement
*Removed LongWrapper (why was it even there?)
*Use foreach loops instead of iterators/for i loops
* Remove local variables that will be returned in the next statement + use ternary operator for some return statements
* Remove some testing code (that was not used any more)
 * Prevent instance creation of the Main class
 * Made some fields private (that should be done consistently in every class)
 * Some more things

+++ src/jasmin/gui/Console.java
@@ -94,7 +94,7 @@ public class Console extends javax.swing.JPanel implements IGuiModule, IListener
 				}
 			}
 		});
-		
+
 		jScrollPane1 = new javax.swing.JScrollPane();
 		setLayout(new java.awt.BorderLayout());
 		jScrollPane1.setViewportView(output);
@@ -156,7 +156,7 @@ public class Console extends javax.swing.JPanel implements IGuiModule, IListener
 		int address_temp = 0;
 		if (newAddress != null) {
 			try {
-				address_temp = new Integer(Parser.hex2dec(newAddress.toUpperCase())).intValue();
+				address_temp = new Integer(Parser.hex2dec(newAddress.toUpperCase()));
 			} catch (Exception ex) {
 				JOptionPane.showMessageDialog(null, "The entered value was not valid!");
 			}
