+++ src/jasmin/gui/SevenSegment.java
@@ -265,7 +265,7 @@ public class SevenSegment extends javax.swing.JPanel implements IGuiModule, ILis
 		int address_temp = 0;
 		if (newAddress != null) {
 			try {
-				address_temp = new Integer(Parser.hex2dec(newAddress.toUpperCase())).intValue();
+				address_temp = new Integer(Parser.hex2dec(newAddress.toUpperCase()));
 			} catch (Exception ex) {
 				JOptionPane.showMessageDialog(null, "The entered value was not valid!");
 			}
@@ -290,7 +290,7 @@ public class SevenSegment extends javax.swing.JPanel implements IGuiModule, ILis
 		int digits_temp = 0;
 		if (newDigits != null) {
 			try {
-				digits_temp = new Integer(newDigits).intValue();
+				digits_temp = new Integer(newDigits);
 			} catch (Exception ex) {
 				JOptionPane.showMessageDialog(null, "The entered value was not valid!");
 			}
