+++ src/jasmin/gui/StripLight.java
@@ -183,7 +183,7 @@ public class StripLight extends javax.swing.JPanel implements IGuiModule, IListe
 		int address_temp = 0;
 		if (newAddress != null) {
 			try {
-				address_temp = new Integer(Parser.hex2dec(newAddress.toUpperCase())).intValue();
+				address_temp = new Integer(Parser.hex2dec(newAddress.toUpperCase()));
 			} catch (Exception ex) {
 				JOptionPane.showMessageDialog(null, "The entered value was not valid!");
 			}
@@ -207,7 +207,7 @@ public class StripLight extends javax.swing.JPanel implements IGuiModule, IListe
 		int digits_temp = 0;
 		if (newDigits != null) {
 			try {
-				digits_temp = new Integer(newDigits).intValue();
+				digits_temp = new Integer(newDigits);
 			} catch (Exception ex) {
 				JOptionPane.showMessageDialog(null, "The entered value was not valid!");
 			}
